1. 删除不必要的common模块 2. 包名/包路径调整
