--- gugu_blog_sync_common/pom.xml
@@ -1,44 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xmlns="http://maven.apache.org/POM/4.0.0"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-    <parent>
-        <groupId>com.gugumin</groupId>
-        <artifactId>gugu_blog_sync</artifactId>
-        <version>0.0.1</version>
-    </parent>
-
-    <artifactId>gugu_blog_sync_common</artifactId>
-
-    <developers>
-        <developer>
-            <name>minmin</name>
-            <url>https://github.com/MinMinGuGu</url>
-        </developer>
-    </developers>
-
-    <scm>
-        <url>https://github.com/MinMinGuGu/GuGu_Blog_Sync</url>
-    </scm>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.projectlombok</groupId>
-            <artifactId>lombok</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.fasterxml.jackson.core</groupId>
-            <artifactId>jackson-databind</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.yaml</groupId>
-            <artifactId>snakeyaml</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.fasterxml.jackson.dataformat</groupId>
-            <artifactId>jackson-dataformat-toml</artifactId>
-        </dependency>
-    </dependencies>
-
-</project>
