+++ gugu_blog_sync_core/src/main/java/com/gugumin/core/event/BaseArticleEvent.java
@@ -1,0 +1,32 @@
+package com.gugumin.core.event;
+
+import com.gugumin.core.pojo.Article;
+import lombok.Getter;
+import org.springframework.context.ApplicationEvent;
+
+import java.util.List;
+
+/**
+ * The type Abs article event.
+ *
+ * @author minmin
+ * @date 2023 /03/26
+ */
+public abstract class BaseArticleEvent extends ApplicationEvent {
+    /**
+     * The Article list.
+     */
+    @Getter
+    protected final List<com.gugumin.core.pojo.Article> articleList;
+
+    /**
+     * Instantiates a new Abs article event.
+     *
+     * @param source      the source
+     * @param articleList the article list
+     */
+    public BaseArticleEvent(Object source, List<Article> articleList) {
+        super(source);
+        this.articleList = articleList;
+    }
+}
