+++ gugu_blog_sync_core/src/main/java/com/gugumin/core/pojo/Article.java
@@ -1,0 +1,43 @@
+package com.gugumin.core.pojo;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+/**
+ * The type Article.
+ *
+ * @author minmin
+ * @date 2023 /03/06
+ */
+@Data
+@AllArgsConstructor
+public class Article {
+    private static final Pattern META_PATTERN = Pattern.compile("^```meta-(json|yaml|properties|toml)");
+    private String name;
+    private String context;
+    private Meta meta;
+    private com.gugumin.core.pojo.MetaType metaType;
+
+    /**
+     * Parse meta from context com . gugumin . core . pojo . meta type.
+     *
+     * @param context the context
+     * @return the com . gugumin . core . pojo . meta type
+     */
+    public static com.gugumin.core.pojo.MetaType parseMetaFromContext(String context) {
+        Matcher matcher = META_PATTERN.matcher(context);
+        if (matcher.find()) {
+            String group = matcher.group();
+            String metaTypeString = group.substring(group.indexOf("-") + 1);
+            for (com.gugumin.core.pojo.MetaType metaType : MetaType.values()) {
+                if (metaType.getValue().equalsIgnoreCase(metaTypeString)) {
+                    return metaType;
+                }
+            }
+        }
+        throw new RuntimeException("无法获取到context中的meta格式");
+    }
+}
