+++ gugu_blog_sync_core/src/main/java/com/gugumin/core/pojo/Meta.java
@@ -1,0 +1,35 @@
+package com.gugumin.core.pojo;
+
+import lombok.Data;
+
+import java.util.List;
+
+/**
+ * The type Meta.
+ *
+ * @author minmin
+ * @date 2023 /03/11
+ */
+@Data
+public class Meta {
+    private List<Category> categories;
+    private List<Tag> tags;
+    private String summary;
+
+    /**
+     * The type Category.
+     */
+    @Data
+    public static class Category {
+        private String parent;
+        private String name;
+    }
+
+    /**
+     * The type Tag.
+     */
+    @Data
+    public static class Tag {
+        private String name;
+    }
+}
