+++ gugu_blog_sync_core/src/main/java/com/gugumin/core/service/IGitService.java
@@ -1,0 +1,24 @@
+package com.gugumin.core.service;
+
+/**
+ * The interface Git service.
+ *
+ * @author minmin
+ * @date 2023 /03/09
+ */
+public interface IGitService {
+    /**
+     * Init repository.
+     */
+    void initRepository();
+
+    /**
+     * Push repository.
+     */
+    void pushRepository();
+
+    /**
+     * Update repository.
+     */
+    void updateRepository();
+}
