+++ gugu_blog_sync_core/src/main/java/com/gugumin/core/service/IHandlerInitSite.java
@@ -1,0 +1,20 @@
+package com.gugumin.core.service;
+
+import com.gugumin.core.pojo.Article;
+
+import java.util.List;
+
+/**
+ * The interface Handler init site.
+ *
+ * @author minmin
+ * @date 2023 /03/26
+ */
+public interface IHandlerInitSite {
+    /**
+     * Gets articles.
+     *
+     * @return the articles
+     */
+    List<Article> getArticles();
+}
