--- gugu_blog_sync_core/src/main/java/com/gugumin/event/DeleteArticleEvent.java
@@ -1,23 +1,0 @@
-package com.gugumin.event;
-
-import com.gugumin.pojo.Article;
-
-import java.util.List;
-
-/**
- * The type Delete article event.
- *
- * @author minmin
- * @date 2023 /03/26
- */
-public class DeleteArticleEvent extends BaseArticleEvent {
-    /**
-     * Instantiates a new Delete article event.
-     *
-     * @param source      the source
-     * @param articleList the article list
-     */
-    public DeleteArticleEvent(Object source, List<Article> articleList) {
-        super(source, articleList);
-    }
-}
