--- gugu_blog_sync_core/src/main/java/com/gugumin/service/IGithubWebhook.java
@@ -1,16 +1,0 @@
-package com.gugumin.service;
-
-/**
- * The interface Github webhook.
- *
- * @author minmin
- * @date 2023 /03/08
- */
-public interface IGithubWebhook {
-    /**
-     * Handler.
-     *
-     * @param payload the payload
-     */
-    void handler(String payload);
-}
