--- gugu_blog_sync_core/src/main/java/com/gugumin/service/IHandlerInitSite.java
@@ -1,20 +1,0 @@
-package com.gugumin.service;
-
-import com.gugumin.pojo.Article;
-
-import java.util.List;
-
-/**
- * The interface Handler init site.
- *
- * @author minmin
- * @date 2023 /03/26
- */
-public interface IHandlerInitSite {
-    /**
-     * Gets articles.
-     *
-     * @return the articles
-     */
-    List<Article> getArticles();
-}
