+++ gugu_blog_sync_core/src/test/java/com/gugumin/event/EventTest.java
@@ -1,6 +1,7 @@
 package com.gugumin.event;
 
 import com.gugumin.CoreTest;
+import com.gugumin.core.event.AddArticleEvent;
 import org.junit.jupiter.api.Test;
 import org.springframework.context.ApplicationEventPublisher;
