--- gugu_blog_sync_starter/src/main/java/com/gugumin/springboot/AutoConfiguration.java
@@ -1,11 +1,0 @@
-package com.gugumin.springboot;
-
-import org.springframework.context.annotation.ComponentScan;
-
-/**
- * @author minmin
- * @date 2023/03/12
- */
-@ComponentScan("com.gugumin")
-public class AutoConfiguration {
-}
