+++ gugu_blog_sync_starter/src/main/java/com/gugumin/starter/AutoConfiguration.java
@@ -1,0 +1,13 @@
+package com.gugumin.starter;
+
+import org.springframework.context.annotation.ComponentScan;
+
+/**
+ * The type Auto configuration.
+ *
+ * @author minmin
+ * @date 2023 /03/12
+ */
+@ComponentScan("com.gugumin.core")
+public class AutoConfiguration {
+}
