+++ gugu_blog_sync_starter/src/main/resources/META-INF/spring.factories
@@ -1,1 +1,1 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.gugumin.springboot.AutoConfiguration
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=com.gugumin.starter.AutoConfiguration
